+++ custom_metric/activity.py
@@ -1,0 +1,9 @@
+import time
+
+from temporalio import activity
+
+
+@activity.defn
+def print_and_sleep():
+    print("In the activity.")
+    time.sleep(1)
