+++ tests/custom_metric/workflow_test.py
@@ -1,33 +1,31 @@
 import uuid
 
 from temporalio import activity
-from temporalio.testing import WorkflowEnvironment
+from temporalio.client import Client
 from temporalio.worker import Worker
 
-from custom_metric.worker import ExecuteActivityWorkflow
+from custom_metric.worker import StartTwoActivitiesWorkflow
 
 _TASK_QUEUE = "custom-metric-task-queue"
 
 activity_counter = 0
 
 
-async def test_sleep_for_days_workflow():
+async def test_custom_metric_workflow(client: Client):
 
     @activity.defn(name="print_message")
     async def print_message_mock():
         global activity_counter
         activity_counter += 1
 
-    async with await WorkflowEnvironment.start_time_skipping() as env:
-
         async with Worker(
-            env.client,
+            client,
             task_queue=_TASK_QUEUE,
-            workflows=[ExecuteActivityWorkflow],
+            workflows=[StartTwoActivitiesWorkflow],
             activities=[print_message_mock],
         ):
-            result = await env.client.execute_workflow(
-                ExecuteActivityWorkflow.run,
+            result = await client.execute_workflow(
+                StartTwoActivitiesWorkflow.run,
                 id=str(uuid.uuid4()),
                 task_queue=_TASK_QUEUE,
             )
